# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EMARA API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      responses:
        "200":
          description: List users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
    post:
      operationId: createUser
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
...
  /projects/{projectId}/stages:
    get:
      operationId: listProjectStages
      tags: [stages]
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Stages
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProjectStage" }
  /stages/{id}:
    patch:
      operationId: updateStage
      tags: [stages]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateStageBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectStage" }
  /projects/{projectId}/quotation:
    get:
      operationId: getProjectQuotation
      tags: [quotations]
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Quotation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuotationOrNull" }
    put:
      operationId: upsertProjectQuotation
      tags: [quotations]
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertQuotationBody" }
      responses:
        "200":
          description: Quotation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quotation" }
  /quotations/{id}/status:
    patch:
      operationId: updateQuotationStatus
      tags: [quotations]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateQuotationStatusBody" }
      responses:
        "200":
          description: Quotation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quotation" }
...
components:
  schemas:
...
    Role:
      type: string
      enum: [admin, sales, designer, draftsman, qs, accountant, client]
...
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        clientId: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, name, email, role, clientId, createdAt]
    CreateUserBody:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        clientId: { type: ["integer", "null"] }
      required: [name, email, password, role]
    UpdateUserBody:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        clientId: { type: ["integer", "null"] }
...
    StageStatus:
      type: string
      enum:
        - locked
        - not_started
        - in_progress
        - waiting_client_approval
        - revision_required
        - approved
        - completed
    ProjectStage:
      type: object
      properties:
        id: { type: integer }
        projectId: { type: integer }
        stageName: { type: string }
        stageKey: { type: string }
        order: { type: integer }
        status: { $ref: "#/components/schemas/StageStatus" }
        assignedToId: { type: ["integer", "null"] }
        assignedToName: { type: ["string", "null"] }
        startDate: { type: ["string", "null"], format: date }
        dueDate: { type: ["string", "null"], format: date }
        completedAt: { type: ["string", "null"], format: date-time }
        requiresClientApproval: { type: boolean }
        notes: { type: ["string", "null"] }
      required:
        - id
        - projectId
        - stageName
        - stageKey
        - order
        - status
        - assignedToId
        - assignedToName
        - startDate
        - dueDate
        - completedAt
        - requiresClientApproval
        - notes
...
    UpdateStageBody:
      type: object
      properties:
        status: { $ref: "#/components/schemas/StageStatus" }
        assignedToId: { type: ["integer", "null"] }
        startDate: { type: ["string", "null"], format: date }
        dueDate: { type: ["string", "null"], format: date }
        notes: { type: ["string", "null"] }
    QuotationStatus:
      type: string
      enum: [draft, sent, approved, rejected]
    Quotation:
      type: object
      properties:
        id: { type: integer }
        projectId: { type: integer }
        areaSqm: { type: number }
        pricePerSqm: { type: number }
        designSubtotal: { type: number }
        shopDrawingPrice: { type: number }
        supervisionPrice: { type: number }
        discount: { type: number }
        tax: { type: number }
        finalTotal: { type: number }
        status: { $ref: "#/components/schemas/QuotationStatus" }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - projectId
        - areaSqm
        - pricePerSqm
        - designSubtotal
        - shopDrawingPrice
        - supervisionPrice
        - discount
        - tax
        - finalTotal
        - status
        - createdAt
    QuotationOrNull:
      oneOf:
        - $ref: "#/components/schemas/Quotation"
        - type: "null"
    UpsertQuotationBody:
      type: object
      properties:
        areaSqm: { type: number }
        pricePerSqm: { type: number }
        shopDrawingPrice: { type: number }
        supervisionPrice: { type: number }
        discount: { type: number }
        tax: { type: number }
      required: [areaSqm, pricePerSqm, shopDrawingPrice, supervisionPrice, discount, tax]
    UpdateQuotationStatusBody:
      type: object
      properties:
        status: { $ref: "#/components/schemas/QuotationStatus" }
      required: [status]
...